Moved slice to list type
